db.cpp: Populate AssociationEventData
So far, the BSSID, MacAddress, HTCapabilites and
VHTCapabilities are filled. ReasonCode is known not
to be available to the controller from the AP
the station is connecting to.

Need to fix current data model - Notifications object
should be part of Controller (Controller.Notifications).

Also editor automatically removed trailing
whitespaces.

PPM-746

Signed-off-by: pablo <pablo@odkq.com>
Signed-off-by: Vladyslav Tupikin <v.tupikin@inango-systems.com>