controller: receive scan request event in dyn_channel_selection_r2_task
The dynamic channel selection r2 task expect a TRIGGER_SINGLE_SCAN event
(with one of the radios mac) to trigger a channel scan.

This trigger will not be implemented until the DCS R2 feature is complete,
in order not to affect current DCS operation.

This commit add an event handler for scan_request events.
Once a scan_request event is received, the request is stored into
the dynamic_channel_selection_r2_task status container.

- Receive TRIGGER_SINGLE_SCAN event in the controller
  dynamic channel selection r2 task.

- Add handler for scan_request events to
  dynamic_channel_selection_r2_task in the controller.
  the handler will store a new scan request into the
  task status container.

PPM-360
---
Signed-off-by: Oren Vormaser <oren.vormaser@devalore.com>