agent: Add radio_on_scan function
On several places in the channel selection task, the task checks if there is a radio that
is in the middle of the background scan. On each of these places that task iterate on
the radios in the database. To prevent code duplication, add a function that does it,
and use it instead.

Signed-off-by: Moran Shoeg <moranx.shoeg@intel.com>