agent: Find the absolute best rank channel
zwdfs_select_best_usable_channel() function on channel selection task is
calculating which channel is the best-ranked channel, and consider the
configured ranking threshold to do it. This means that if there is a
better-ranked channel, but it is not passing the ranking threshold,
it will not be chosen.
For testing and debugging purposes, also calculate the absolute
best-ranked channel and print a message if it didn't get selected.

On the way, rename "channel_selection" varaible to "selected_channel".

Signed-off-by: Moran Shoeg <moranx.shoeg@intel.com>