controller:db: change parameter to const reference
Change parameter 'path_to_object' of 'dm_set_ht_capabilities()' method
to the const reference as we want it to be used later on outside the method
and we don't want this parameter to be modified inside of our method.
Remove unnecessary dote from log message.
Changed comment on line 1529 in db.cpp from "To do:" to "TODO:".

PPM-515
--
Signed-off-by: Krystyna Oliinyk <k.oliinyk@inango-systems.com>