bcl: lambda uses unique_ptr after move
Lambda function is capturing a unique_ptr local variable by reference
and after that, the pointer is moved to a map. If the lambda code was
ever executed, the value of captured pointer would be nullptr.

More over, that code is never used, so the simplest way to fix the
problem is remove it.

Since the code to roll back actions in case of failure is used in one
step only, explicitly undo that step if error and remove the whole
roll back construct.

Signed-off-by: Mario Maz <mmazzapater@gmail.com>