documentation: README.md: remove obsolete instructions
repo manifest is not used now for fetch prplMesh sources.
List of build requirements is outdated.

Set in README.md section to use git clone for that.

Add link to Dockerfile of image used on CI. It'll contain
actual list of Ubuntu packages to install for prplMesh
build.

Split build section into "Build in Docker" and "Native Build".

Signed-off-by: Frederik Van Bogaert <frederik.vanbogaert@mind.be>
Signed-off-by: Ivan Efimov <i.efimov@inango-systems.com>