bpl: Handle non-numeric value
For ZWDFS to work, the interface channel shall be defined on UCI as "first_non_dfs".
This string causing BPL error, since the stoi function fails to convert it to an integer.
This results in putting zero as a default value and error print in the log.
The function which read the channel is being polled from the platform manager,
so the error print floods the log file.
Also, the main usage for the channel reading it to tell the ap_manager if it is on ACS
mode or not, so the default 0 value is totally wrong.

To solve this, in case of any non-numeric value is found on the channel field
(except for "auto), the function will assign the channel to 0.

Signed-off-by: Moran Shoeg <moranx.shoeg@intel.com>