ci: preserve env variables for sudo commands
Certification tests execute some commands in elevated mode
(using the sudo command). These commands may rely on environment
variables which are set by previous commands or by the gitlab runner
itself.

Use the "-E" argument of the sudo command to preserve the existing
environment variables.

This change is introduced to support an update that was made to the
"power_cycle.py" script, to support the APC used in the ISR testbed.
The updated script is checking the environment for a custom
configuraiton file. The ISR gitlab runner sets this variable
pointing to the ISR specific configuration file.

Signed-off-by: Vitaly Bukhovsky <vitaly@devalore.com>