tree-wide: add print at the beginning of processes start
In some platforms, we encountered a scenario where although the call to
start the prplmesh processes is called, the processes are not started.

To identify if the issue is due to some early crash of the process, or
the platform not starting the process at all we added a print at the
beginning of each process.

Signed-off-by: Adam Dov <adamx.dov@intel.com>