agent: store incoming scan results in AgentDB
When the Channel-Scan-Result-Notification is handled by the agent, the
result needs to be cached to the AgentDB, as well as forwarded to the
controller to not disrupt the current Channel-Scan feature.

Later when the EasyMesh R2 Channel-Scan Task is ready to handle the
incoming events, this code will be handled directly in the task.

- Store incoming DCS results (Scanned Neighboring AP) to the AgentDB.
PPM-361
Signed-off-by: Itay Elenzweig <itayx.elenzweig@intel.com>