agent: link_metrics_collection_task: actually emit link metrics
The link metrics are recorded in sInterfacePairInfo structures that are
embedded in the tlvTransmitter/ReceiverLinkMetric objects. However, when
setting the values in the sInterfacePairInfo structures, the object was
taken by value instead of by reference, so the values were never updated
in the tlvTransmitter/ReceiverLinkMetric object itself. As a result, the
data in the TLV was just random (whatever used to be in the buffer
before, usually 0).

Fix this by taking the sInterfacePairInfo structures by reference.

Reported-by: Yuriy Masechko <yuriy.masechko@globallogic.com>
Signed-off-by: Arnout Vandecappelle (Essensium/Mind) <arnout@mind.be>