agent: backhaul_manager: replace thread::work() with even_loop::run()
https://jira.prplfoundation.org/browse/PPM-753

This commit replaces the use of `Socket` and `socket_thread` classes
with `EventLoop` class and the new sockets API. Next commit will remove
all references to these deprecated classes and unused methods from the
backhaul manager class.

In main, replace call to `backhaul_mgr.start()` with a call to the
temporarily named method `backhaul_mgr.to_be_renamed_to_start()`. This
means that backhaul manager is no longer started as a thread but it
will block on event loop for I/O operations in main thread.

As a consequence of not being started as a thread, the UDS path passed
to `transport_socket_thread` and in turn to `socket_thread` is not used
any more. So, since now there is no conflict, remove the "_temp" suffix
when `backhaul_manager_uds_path` is defined so the right UDS path is
used in the CMDU server.

Uncomment the lines that create the TCP server to receive certification
commands from now on using the UCC listener, which from now on will not
be a thread any more either.

In backhaul manager, send CMDU messages to slaves using the CMDU server
and to broker server using broker client; replace Socket classes used
to receive events from hostapd control interface with file descriptors;
and use the CMDU client factory to create the client to connect to the
CMDU server on platform manager.

Signed-off-by: Mario Maz <mmazzapater@gmail.com>