nl80211: disable VAPs by using start_disabled and removing ssid
Commenting out the VAPs to disable is problematic because when issuing
the UPDATE command on openwrt, hostapd will be restarted, leading us
to a slave reset.

Since we anyway assume the VAPs to always be there on start, instead
of commenting out the VAPs use "start_disabled".

As noted in the commit message of
"df9b594ca375ae48ef4d7653bef047c6ddc5cb08", the backhaul manager
relies on the ssid parameter being absent from the configuration. This
is already how we disable a VAP in the dwpal implementation: set
start_disabled and remove the ssid parameter. Do the same for the
nl80211 implementation.

The following changes are introduced:

- Introduce 'disable_vap()'. It both sets 'start_disabled' to 1, and
remove the ssid parameter.

- We don't need to comment and uncomment VAPs, so remove the 2
corresponding functions.

- Also remove the parts of the tests that use
'uncomment_vap()'. Replace the ones that use 'comment_vap()' with
"disable_vap".

- Use disable_vap() when a vap needs to be disabled. Also log the
operation to the DEBUG logs.

- At the end of the configuration of a VAP, we don't need to uncomment
it anymore, so just remove 'start_disabled'.

Fixes https://jira.prplfoundation.org/browse/PPM-803

This will should allow us to pass the following tests on the Turris Omnia:
4.5.1
4.5.2
4.6.1

4.7.9 was also affected by this issue, but it will now fail because of
the following issue:
https://jira.prplfoundation.org/browse/PPM-850

Signed-off-by: Raphaël Mélotte <raphael.melotte@mind.be>