bpl: cfg: move methods that write to UCI to common
https://jira.prplfoundation.org/browse/PPM-832

This is a preparatory commit.

We need to persist wireless credentials over reboots and this is to be
implemented in bpl_cfg.

In the UCI implementation, code for writing to UCI already exists in
bpl_db. It needs to be refactored so it can be called from bpl_cfg as
well.

Move existing methods that write to UCI to a common directory so they
can be reused both from bpl_db and bpl_cfg.

Signed-off-by: Mario Maz <mmazzapater@gmail.com>