tlvf: add reporting_detaul to sBeaconRequest11k
The 1905.1 Beacon Metrics Query message includes a "reporting detail"
field. The VS specific message we map it to ("sBeaconRequest11k")
doesn't include such a field.

It includes "report" and "request", but they really seem unrelated.

As a consequence, when we send a Beacon Request, we currently hardcode
this value to 2, while what we should do is to use the value sent in
the Beacon Metrics Query. See the corresponding line in the Multi-AP
standard:
"The Operating Class, Channel Number, BSSID and Reporting Detail
fields in the Beacon Request shall be set to the corresponding values
specified in the query"

Add "reporting_detail" in "sBeaconRequest11k", it will be used in a
follow-up commit to set the correct value in the Beacon Request.

Signed-off-by: Raphaël Mélotte <raphael.melotte@mind.be>