agent: use the reporting_detail value from the Beacon Metrics Query
When we send a Beacon Request, we're supposed to use the
reporting_details value from the Beacon Metrics Query.

Use the new value we introduced in SBeaconRequest11k to set it.

Note that it's only set in the dwpal implementation for now, because
the nl80211 implementation is missing additional fields in the Beacon
request.

A follow-up commit will also use the reporting_detail value for the
nl80211 implementation.

Signed-off-by: Raphaël Mélotte <raphael.melotte@mind.be>