common: save the operating class in channel reports
When we receive a Beacon Metrics Query, each Channel Report includes
an operating class.

In the dwpal BWL, it's not needed because hostapd sets it by itself.

However, the for nl80211 implementation we need to save and set the
operating class.

Since the 1905.1 TLV saves the operating class as the first element of
'ap_ch_report', do the same in 'SBeaconRequest11k'.

Also adapt the dwpal BWL to skip the operating class since it's not
needed.

A follow-up commit will use the operating class in the nl80211
implementation.

Signed-off-by: Raphaël Mélotte <raphael.melotte@mind.be>