common: ucc_listener: add validate_program_parameter
For the R2 certification, the UCC uses a new value for "program":
"mapr2".

Add "supported_programs", which holds the values we support. Since
"mapr2" is one of them, add it.

Add validate_program_parameter(), a helper to check if the "program"
parameter is valid, and log any error. It will be used in a follow-up
commit.

Signed-off-by: Raphaël Mélotte <raphael.melotte@mind.be>