agent: backhaul: remove wired interface in wireless flow
Despite the fact that wired interface is no longer added in
`dev_reset_deafult` we still want to delete it in a wireless flow.

Note:
We will need to have another way to communicate with the device
(e.g. a network interface outside of the bridge or a serial cable)
otherwise removing the interface that is used to send commands
will make the router unreachable.

PPM-838

Signed-off-by: Vitalii Komisarenko <vitalii.komisarenko@gmail.com>