controller: db: get_hostap_vap_list: don't return reference to temporary
In the error paths of get_hostap_vap_list, the following construct is
used:

    return *(std::make_shared<...>());

This creates a shared pointer temporary that gets destroyed at the end
of the function. Since that was the only reference to this shared
pointer, the underlying object will be destroyed as well.

This was not detected up to now because we normally don't end up in the
error path, and anyway there are almost no new allocations until the
return value gets used, so the freed memory references are unlikely to
cause problems.

Fix this by returning a static invalid_vap_list object instead.

This solution is not perfect, because the return value is non-const.
Fortunately, it is only used non-const in one place: db::add_vap(). In
that place, there's a preceding check that makes sure we don't end up in
one of the error branches.

Delay improving on this approach until we can spend time on a proper
refactoring of the controller db, introducing nodes for VAPs as well.

Fixes https://jira.prplfoundation.org/browse/PPM-857

--
Signed-off-by: Arnout Vandecappelle (Essensium/Mind) <arnout@mind.be>