bpl: Remove bpl flooding print
When bpl tries to read a parameter that is not on the database, it prints an error message but does not return failure.
Commit fa8818b has added new bpl reading to cfg_get_wifi_params.
The problem is that this function is being called each second, and on platform
where some uci parameter is missing, an error print will flood the console log.

To fix that, remove the error print.

Signed-off-by: Moran Shoeg <moranx.shoeg@intel.com>