PPM-858 Resolve prplMesh processes hang under high loads
This MR handles various issues that were observed on the Axepoint
platform under heavy loads.
Specifically when connecting 64+ STAs to a single platform.

Large number of STAs generate a lot of control traffic between
the Controller and the Agent (e.g. monitoring requests/responses),
which can cause starvation for various processed within the prplMesh processes.

This MR handles these issues by enforcing the following rules/limitations:

Set a minimal timeout for TX operations on internal sockets inside
the transport process.
This prevents the transport from blocking when one of the receiving processes
do not clean their RX buffers.
Limit the task pools in the controller and the agent for processing duration
and the number of tasks that can be handled in each slot.

Closes PPM-858