ci: retry downloading artifacts and fetching git remote
The latest nightly test on the Turris Omnia failed on fetching the
changes:
https://gitlab.com/prpl-foundation/prplmesh/prplMesh/-/jobs/873189718

It turns out there are settings to retry fetching the
changes and downloading the artifacts, but they default to 1.

Set them to 3 to try to reduce the likelihood of having tests failing
because of temporary network issues.

The variables are documented here:
https://docs.gitlab.com/ee/ci/runners/#job-stages-attempts

Signed-off-by: Raphaël Mélotte <raphael.melotte@mind.be>