openwrt: reconfigure wps on reload
Update prplWrt to include a new hostapd patch to reconfigure the wps
settings on reload.

This will first be merged in the prplMesh target
branch ('prplwrt-1.0-rc1-prplmesh') because it's definitely not worse
than the previous patch that introduced the same functionality, and
because we need it to be able to pass any test where we use WPS on one
of our backhaul APs.

At the same time, the patch will be submitted for a review between the
prplWrt/prplMesh developers, and if no issue is raised it will be sent
to upstream hostapd.

Signed-off-by: Raphaël Mélotte <raphael.melotte@mind.be>