agent: ap_manager: report empty association frames
Even when sACTION_APMANAGER_CLIENT_ASSOCIATED_NOTIFICATION contained
no association frame, no debug message was printed and an empty
association frame was added to the notification. As a result, client
capability reports also contained extraneous message data at the end
corresponding to the length of the association frame.

This is because the '!msg->params.association_frame' won't check for
an empty association frame.

To fix it, check if the association frame length.

Signed-off-by: Raphaël Mélotte <raphael.melotte@mind.be>