config: don't use the Qualcomm device as wireless onboarding agent
In some wireless backhaul tests, we don't get any reply to our M1
message when the qualcomm agent is used.

After a few tests (see [1]) it looks like the combination that works
best is to use the Mediatek device as the agent we onboard to, and the
Broadcom device as a controller.

The list in [2] reports the test cases where this issue occurs. For
those tests, replace the Qualcomm agent by the Mediatek agent, and
use the Broadcom device as controller.

The following tests were changed:
- MAP-4.2.3_BHWIFI
- MAP-4.4.2_BHWIFI
- MAP-4.4.3_BHWIFI
- MAP-4.5.3_BHWIFI
- MAP-4.7.1 (wifi BH only)
- MAP-4.7.2 (wifi BH only)
- MAP-4.7.5_BHWIFI
- MAP-4.7.6_BHWIFI
- MAP-4.7.9_ETH
- MAP-4.8.4_BHWIFI
- MAP-4.8.5_BHWIFI
- MAP-4.10.5 (wifi BH only)

[1]: https://gitlab.com/prpl-foundation/prplmesh/prplMesh/-/pipelines/223350077
[2]: https://jira.prplfoundation.org/browse/PPM-887?focusedCommentId=14310&page=com.atlassian.jira.plugin.system.issuetabpanels%3Acomment-tabpanel#comment-14310

Signed-off-by: Raphaël Mélotte <raphael.melotte@mind.be>