TLVF: Change sNodeBssSteerTarget reason to int16_t
This is preperative commit.
Profile1 tlv doesn't add the reason attribute. Since this struct is
common for any steering request, we will need to mark the reason as
invalid on profile1 steering request.

Change the reason from uint8_t to int16_t to be able to mark a negative
value.

PPM-860

Signed-off-by: Semyon Barenboym <semyon.barenboym@devalore.com>