tlvf: CmduMessageRx::parseNextTlv: add missing TLV types
Based on the TLV types defined in eTlvType.h and eTlvTypeMap.h, complete
the big switch in the parseNextTlv function with all the missing TLVs.

TLV_UNASSOCIATED_STA_LINK_METRICS_QUERY and
TLV_UNASSOCIATED_STA_LINK_METRICS_RESPONSE are defined in eTlvTypeMap,
but no corresponding classes exist at the moment. For those, issue a
debug message and generate a tlvUnknown.

With this, the switch covers all cases so we can replace the default
case with a fatal log.

This way, the compiler will give a warning if an enum value is added but
parseNextTlv is not updated.

Fixes: https://jira.prplfoundation.org/browse/PPM-673

Signed-off-by: Arnout Vandecappelle (Essensium/Mind) <arnout@mind.be>