tlvf: CmduMessageRx::parseNextTlv: use symbols for TLV types
parseNextTlv is a big switch that matches TLV type to the corresponding
class. Currently it uses decimal literals in this mapping. It's better
to use the symbolic constants defined in eTlvType.h and eTlvTypeMap.h
instead.

This will make it a bit easier to verify that it's correct, and also
easier to insert new ones.

Using the symbolic constant requires us to convert the uint8_t value
into an enum first. Since we have two enums (eTlvType and eTlvTypeMap),
split the switch in two, one for each enum, and put each of them in a
separate function. The automatically generated eTlvTypeValidate::check
function can be used to find which enum has to be used.

For the eTlvType enum, all values are covered so we don't need a default
case in the switch. The compiler will tell us when a value is added and
we forget to extend the switch.

For the eTlvTypeMap enum, not all values are covered yet, so the default
case is kept.

Also promote the log message in parseNextTlv itself to INFO. An unknown
TLV is an unexpected situation, so it makes sense to lo it more
aggressively.

Signed-off-by: Arnout Vandecappelle (Essensium/Mind) <arnout@mind.be>