tlvf.py: make python3 the default
Python2 is deprecated, and some distributions (e.g. alpine) already
removed it from their stable release.

Running '2to3' 3.7.3 on it revealed only irrelevant things (like
iterating over dict.items() - 2to3 suggests converting to a list which
is not useful in this case) except for the unicode string which is
changed here.

We also need to adapt the cmake configuration to require
python3. Unfortunately FindPythonInterp cannot be used to find python3
when python2 is available (see [1]). Instead, we need to use
FindPython3, which was only added in cmake 3.12. The cmake version we
have in the Ubuntu Bionic builder is '3.10.2'.  To solve this, use
"command -v" directly to resolve the path to the python3
interpreter. This may not work on Windows setups, but we anyway
dropped Windows support a while ago.

This will allow us to remove the python2 dependency in our builders.

[1]: https://cmake.org/cmake/help/latest/module/FindPythonInterp.html

Signed-off-by: Raphaël Mélotte <raphael.melotte@mind.be>