CMakeLists: allow override RDKB related CMake variables from shell
There is overriding CMake variables in main CMakeLists.txt file
for case of building prplMesh for RDKB.
That's done such way that not allow redefine that variables from
shell.

Fpr example, backhaul iface and bridge name are set for Puma7 and
not if build for example for another board then it'll be needed
to set another values from package/recipe.

Define needed variables as cache variable.

Signed-off-by: Ivan Efimov <i.efimov@inango-systems.com>