tlvf: replace unicode dash character
The character that was used as a dash was the unicode 0x2013. Replace
it with an ascii dash.

Without this, compiling with python3 will fail with:
UnicodeDecodeError: 'ascii' codec can't decode byte 0xe2 in position 1695: ordinal not in range(128)

This happened when doing this with python3:
yaml_inst = yaml.load(stream, Loader=OrderedDictYAMLLoader)

Signed-off-by: Raphaël Mélotte <raphael.melotte@mind.be>