bwl: nl80211: remove optional subelement
https://jira.prplfoundation.org/browse/PPM-860

In BSS_TM_REQ command, the format of the neighbor field is:
neighbor=<BSSID>,<BSSID Information>,<Operating Class>,<Channel
Number>,<PHY Type>[,<hexdump of Optional Subelements>]

Currently we are setting the last optional sub-field to 255 and
BSS_TM_REQ command fails in Turris-Omnia and Gl.iNet devices. The reason
is "255" is not a valid ASCII hex string. Moreover, Optional Subelements
are TLVs so at least 3 bytes shall be specified.

In BWL::DWPAL, the 255 is Neighbor preference and not Optional
Subelements. It is later translated to 0301ff when filling in the
Optional Subelements sub-field.

So, to fix the problem, remove this optional and useless field from
BWL::NL80211.

Signed-off-by: Mario Maz <mmazzapater@gmail.com>