agent: backhaul_manager: move creation of tasks to constructor
https://jira.prplfoundation.org/browse/PPM-753

Modify backhaul_manager_thread to create worker tasks in constructor
instead of in the `init()` method.

Worker tasks must be created only once so it is safer if they are
created in the class constructor instead of in a method that can be
called more than once. Also, if something goes wrong while creating
the tasks, application should log a fatal error and abort.

Note that method `TaskPool::add_task()` takes care of checking if
provided task pointer is nullptr and aborts if so.

This is a preparatory commit. Next commits will replace `init()`
with a `start()` method and will add a `stop()` method to clean-up
allocated resources.

Signed-off-by: Mario Maz <mmazzapater@gmail.com>