ap-manager: defragment generate connected clients events
As part of the init sequence, the AP manager generates client
association events for any pre-connected client. The generation is
triggered on salve request of:
ACTION_APMANAGER_HOSTAP_GENERATE_CLIENT_ASSOCIATION_NOTIFICATIONS_REQUEST.

The current implementation creates a burst of events to the slave which
in turn, creates a burst of events to the controller and in the case of
many clients connected can lead to full buffers and loss of messages
(due to the fix in transport that resolves the deadlock by adding a
timeout to the write operations).

Instead, the above message will set a flag of "generate connected
clients events need to be done" when the request above-mentioned is
received and that flag will be checked as part of the next
after-select() of the thread.
When the generation of events for all pre-connected clients is complete
the flag is set to false and will not change until the next time the
request is received (currently only after slave join).

The generate_connected_clients_events() is adjusted to defragment the
generation of the event and to statically save "last point we stopped"
and continue from that point on the next call to the function.
The function now returns true on completion or false on error or if we
still have clients we need to generate events for.

Signed-off-by: Adam Dov <adamx.dov@intel.com>