tests: real device: restart networking on deploying prplmesh
Backhaul manager removes the wired interface from the bridge in
the wireless flow. Restarting networking reconfigures the bridge
so that the interface is put back.

It is already done AFTER each test. But if boardfarm is killed
in the middle of the test (e.g. due to timeout or out of memory)
the bridge may lack the wired backhaul interface.

To address this issue this commit restarts networking in
the deploy script to have the bridge in the desired state before
running the very first test.

PPM-838

Signed-off-by: Vitalii Komisarenko <vitalii.komisarenko@gmail.com>