mon_wlan_hal_dwpal: generate event only for associated clients
When an unassociated client is listed in the hostapd STAs we need to
skip it and not generate a client associated event for it.

The way we read the information from the hostapd requires us to save
the client mac before skipping it.

Removed the 'return nullptr' in the generate_client_assoc_event() in
case of the not associated client and the function will parse the client
mac and other information.

In the calling function generate_connected_clients_events() update the
last-handled client mac and skip sending the event to the monitor if
the returned result is set to SKIP_CLIENT_NOT_ASSOCIATED.

Signed-off-by: Adam Dov <adamx.dov@intel.com>