monitor: defragment generate connected clients events
As part of the init sequence, the monitor thread adds to its DB any
pre-connected client. In the case where the number of clients is large,
this process can take several seconds that means the monitor doesn't
handle messages and events for several seconds, while any request or
event is unhandled during that time.

In the new implementation, a new flag which means "generate connected
clients events is needed" is set to true and on the following
after-select on the steady-state the monitor will generate the events
for the pre-connected clients (in small fragments not to cause a long
time of non-responsiveness).

When the generation of events for all pre-connected clients is complete
the flag is set to false and will not change until the next time the
monitor initialization.

The generate_connected_clients_events() is adjusted to defragment the
generation of the event and to statically save "last point we stopped"
and continue from that point on the next call.
The function now returns true on completion or false on error or if we
still have clients we need to generate events for.

Signed-off-by: Adam Dov <adamx.dov@intel.com>