son_slave: add pre-operational state
It is a better practice not to perform unneeded repeated operations on
each thread wake-up, specifically unneeded use of blocking code-segments
like using the agent DB.

The son-slaves are required to set the stop_on_failure_attempts from the
agent DB value on the transition to the operational state. Since we
don't have a state-entry functionality (to perform on each transition
to this state), the operation was implemented inside the operational
state itself even though this should be performed once.

This may cause the thread to block on Agent DB event though there isn't
a real need for it.

Added STATE_PRE_OPERATIONAL state.
All locations that change state to STATE_OPERATIONAL are changed to move
to PRE_OPERATIONAL state.
After setting the value of stop_on_failure_attempts from agent DB the
state is changed to OPERATIONAL.
Removed the unneeded set of stop_on_failure_attempts from
STATE_OPERATIONAL.

Signed-off-by: Adam Dov <adamx.dov@intel.com>