backhaul_manager: add pre-operational state
It is a better practice not to perform unneeded repeated operations on
each thread wake-up, specifically unneeded use of blocking code-segments
like using the agent DB.

The backhaul-manager is required to finalize any required slaved on
the transition to the STATE_OPERATIONAL state, as well as set the
pending_enable value to false if the connection_type is set to
eConnectionType::Invalid.

Since we don't have a state-entry functionality (to perform on each
transition to this state), the operation was implemented inside the
operational state itself even though this should be performed once.

This may cause the thread to block on Agent DB event though there isn't
a real need for it.

Added STATE_PRE_OPERATIONAL state.
All locations that change state to STATE_OPERATIONAL are changed to move
to PRE_OPERATIONAL state.
After setting the values from agent DB the state is changed to
the STATE_OPERATIONAL state.
Removed the unneeded operations from the STATE_OPERATIONAL state.

Signed-off-by: Adam Dov <adamx.dov@intel.com>