tests: boardfarm: request link metrics for specific STA
https://jira.prplfoundation.org/browse/PPM-771

The CombinedInfraMetrics test is failing always, not randomly.
As part of this test, a Link Metric Query message is sent to agent 1
asking for link metrics for a specific neighbor: the controller.
However, the agent does not consider the controller a neighbor because
the agent has not received a Topology Discovery message from the
controller yet. As a consequence, the agent responds with a Link Metric
Response message with an invalid neighbor result code and thus the test
fails.

There are several possible solutions:
- Modify the test so it doesn't ask for link metrics for a specific
  neighbor but for all known neighbors.
- Modify the test so a Topology Discovery message is sent from the
  controller before sending the Link Metric Query message.
- Modify the test so it asks for link metrics for another neighbor: the
  associated station.

We pick the last option as it is the easiest to implement and it does
not modify the nature or scope of the test.

Signed-off-by: Mario Maz <mmazzapater@gmail.com>