bpl: cfg: create method to get config file name.
https://jira.prplfoundation.org/browse/PPM-832

Configuration file to use is the first file found in an ordered list
of possible candidates.

Create a method to get the name of the configuration file to use and
refactor `cfg_get_param()` to call the new method.

This refactoring is intended to avoid code duplication, since a
follow-up commit will also use this new method to save a parameter
into the configuration file.

Signed-off-by: Mario Maz <mmazzapater@gmail.com>