bpl: cfg: create method to get all config parameters
https://jira.prplfoundation.org/browse/PPM-832

Create a method to get the value of all parameters in configuration
file for which name a given predicate evaluates to true.

Refactor `cfg_get_param()` to call the new method.

This refactoring is intended to avoid code duplication, since a
follow-up commit will also use this new method to read all parameters
which name starts with a given prefix.

Signed-off-by: Mario Maz <mmazzapater@gmail.com>