platform_manager: store the config with a reference
The main slave has a comment saying:
// The platform manager updates the beerocks_slave_conf.sta_iface in the init stage

However, the platform manager takes the beerocks_slave_conf as a const
reference, so there is no way it could have modified it.

In fact, only a copy of the configuration was stored in the platform
manager, so any change to it was not visible in the main slave.

It looks like this configuration is only used in the platform manager
at initialization. Since the initialization happens right after the
platform manager is created, it should be safe to assume the
configuration object will stay alive for much longer after the
platform manager is initialized.

Store a non-const reference to the configuration. We also need to drop
its constness in the PlatformManager constructor.

Fixes PPM-587

Signed-off-by: Raphaël Mélotte <raphael.melotte@mind.be>