openwrt: build.sh: capture exit code of docker build pipeline
Since commit 2df60e4d46, the 'docker build' command is executed as the
first command of a pipeline. Unfortunately, POSIX defines the exit code
of a pipeline as the exit code of the *last* command of the pipeline.
Thus, we don't capture the exit code of the 'docker build' command, and
a failure will simply be ignored.

If a `-t` option was given, and this tag was not yet in use, the problem
will be discovered on the next 'docker exec' command, since the image
doesn't exist if the build failed. However, if an image with the same
tag already exists (which will typically be the case if the user builds
locally), then the error will go unnoticed.

Fix this by setting the pipefail option. This causes the exit code to
come from the last *failing* command of the pipeline. Thus, the exit
code of 'docker build' will be captured.

pipefail is a bashism, so specify bash as the shell.

Signed-off-by: Arnout Vandecappelle (Essensium/Mind) <arnout@mind.be>