transport: handle_packet: only forward IEEE1905.1 messages
Since the big refactoring of the transport, both IEEE1905.1 and LLDP
messages are handled by Ieee1905Transport::handle_packet(). However, the
flow specified there is only valid for IEEE1905.1 packets - other than
the first step that verifies the packet.

This mostly works because update_neighbours and de_duplicate_packet
check that it's 1905.1 and return if not. However, forward_packet
doesn't have such a check, and it really applies only to IEEE1905.1.
The result is that LLDP packets get send to the agent, which tries to
parse them as 1905.1 packets (because the agent doesn't check type
type).

To solve this:

- only try to forward 1905.1 packets;
- for good measure, only do the update_neighbours, de_duplicate and
  forward part of the flow for 1905.1 packets.

Also convert the debug log in the three functions into a warning, and add
the missing warning log in de_duplicate_packet.

Also remove EOL whitespace in comments (automatic by editor).

Fixes: https://jira.prplfoundation.org/browse/PPM-302

Signed-off-by: Arnout Vandecappelle (Essensium/Mind) <arnout@mind.be>