switch channel task: switch-channel-as-a-service
Switch Channel As a Service introduces the ability
to request a switch channel and wait
for the switch channel to happen without taking care for the
entire underlying mess of different driver messaging, timeouts
CAC requirments, etc.

A client only needs to:
* request a switch channel
* wait for switch channel report
The undelying work is done by the switch channel task

The task has internal fsm-per-interface.
Messages arived to the task are processed and
sent to the correct fsm.

Except from handling user requests for
switching a channel, the task also handles
these events initiated by the driver:
* csa notification
* csa error notification
* cac started
* cac complited

The task takes care for handling these events,
manage timeouts and report about the result
of the switch channel.

Read more in
https://confluence.prplfoundation.org/display/PRPLMESH/Switch+Channel+Task

PPM-867

Signed-off-by: Ran Regev <ran.regev@devalore.com>