MAP-4.8.2: always use fronthaul SSID
This test incorrectly tries to associate the CTT STA with the backhaul
BSS. Since we reject a STA on the backhaul BSS if it doesn't have the
Multi-AP Extension subelemnt with the bSTA bit set, this fails on our
devices.

Fix this by using Multi-AP-1 as SSID, regardless of whether multiple
BSSes are supported or not.

Update the easymesh_cert repository with the above change.

Fixes: https://jira.prplfoundation.org/browse/PPM-679

Signed-off-by: Arnout Vandecappelle (Essensium/Mind) <arnout@mind.be>