tools/deploy_firmware.py: remove wireless config before update
Different versions of the wave6 driver have different expectations of
the content of the wireless config. The "correct" config is distributed
in base_files, with an additional run-once script to set the MAC
addresses.

To make sure a correct wireless config is on the device when an upgrade
is done:
- remove the existing /etc/config/wireless from the overlay;
- also remove /etc/uci-defaults/15_wireless-generate-macaddr - this is
  an empty file to indicate that the script from rom is deleted; by
  removing the empty file, the script will run again on next boot.

This is only done for NetgearRax40-style devices. Upstream openwrt
normally doesn't have such incompatible changes.

An alternative would be to remove the entire overlay when an update is
done. This is probably better. However, all the devices in the various
test setups do have some configuration that has to be retained. For
example, in the certification testbed, the devices must have specific IP
addresses, and a specific bridge setup that is different from the
default. So the correct configuration would have to be applied to the
device after the update. This approach would actually be better, because
it gives us much better control of the configuration of each device.
However, it's a lot more work, so we chose the easy path for now.

Fixes: https://jira.prplfoundation.org/browse/PPM-911

Signed-off-by: Arnout Vandecappelle (Essensium/Mind) <arnout@mind.be>