bwl: Add hybrid_mode_supported function
Add the function "hybrid_mode_supported" to all bwl types.
This function says if the radio can be configured as a hybrid
mode (e.g fBSS=1 & bBSS=1).
This is needed for reporting correctly the "combined front and back"
field on the Radio Advanced Capabilities TLV that is required for the
traffic separation flow.

Currently implemented as a stub.

Signed-off-by: Moran Shoeg <moranx.shoeg@intel.com>